LayoutTests/http/tests/referrer-policy-iframe/no-referrer/same-origin.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of no-referrer referrer policy when same origin.");
jsTestIsAsync = true;

if (window.internals)
    internals.setTrackingPreventionEnabled(false);

window.onmessage = function(event) {
    referrer = event.data.referrer;
    // Should be the empty string
    shouldBeEqualToString("referrer", "");
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    finishJSTest();
}
</script>
<iframe src="http://127.0.0.1:8000/referrer-policy/resources/document.html" referrerpolicy="no-referrer"></iframe>
</body>
</html>
